move helper functions to CalendarUtils, improve code readability
