+++ src/main/java/com/google/sps/server/ScheduleCalendar.java
@@ -22,10 +22,7 @@ import com.google.api.services.calendar.model.EventDateTime;
 import com.google.api.client.util.DateTime;
 
 import java.util.Date;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Collection;
-import java.util.Iterator;
 import java.util.Arrays;
 import java.io.IOException;
 
@@ -71,18 +68,16 @@ public class ScheduleCalendar {
     try {
       // get current date, so can add the courses to the appropriate days of the current month 
       java.util.Calendar calendar = java.util.Calendar.getInstance();
-      int nextSun = getDateOfNextSunday(calendar);
-      int month = getCurrentMonth(calendar);
-      int year = getCurrentYear(calendar);
+      int nextSun = CalendarUtils.getDateOfNextSunday(calendar);
+      int month = CalendarUtils.getCurrentMonth(calendar);
+      int year = CalendarUtils.getCurrentYear(calendar);
 
       // for each course in the schedule, add an event to the calendar
-      Iterator<Course> courseIterator = schedule.getCourses().iterator();
-      while (courseIterator.hasNext()) {
-        Course currentCourse = courseIterator.next();
+      for (Course currentCourse: schedule.getCourses()) {
         List<TimeRange> sectionTimes = currentCourse.getSections().get(0).getMeetingTimes();
         for (int i = 0; i < sectionTimes.size(); i++) {
-          String startTime = calculateStartTime(sectionTimes.get(i), month, year, nextSun);
-          String endTime = calculateEndTime(sectionTimes.get(i), month, year, nextSun);
+          String startTime = CalendarUtils.calculateStartTime(sectionTimes.get(i), month, year, nextSun);
+          String endTime = CalendarUtils.calculateEndTime(sectionTimes.get(i), month, year, nextSun);
           addEvent(currentCourse, startTime, endTime);
         }
       } 
@@ -101,100 +96,6 @@ public class ScheduleCalendar {
   }
 
    /**
-   * Helper function to calculate the day of the month of the next Sunday,
-   * to know when to start the schedule. 
-   *
-   * @param calendar an instance of Calendar, so can know the current date
-   * @return an int representing the date of next Sunday
-   */
-  public static int getDateOfNextSunday(java.util.Calendar calendar) {
-    int currentDay = calendar.get(java.util.Calendar.DAY_OF_MONTH);
-    int currentDayOfWeek = calendar.get(java.util.Calendar.DAY_OF_WEEK);
-    int nextSun = currentDay + 8 - currentDayOfWeek;
-    return nextSun;
-  }
- 
-   /**
-   * Helper function to know the current month. 
-   *
-   * @param calendar an instance of Calendar, so can know the current date
-   * @return an int representing the month, adjusted for DateTime's formatting
-   */
-  public static int getCurrentMonth(java.util.Calendar calendar) {
-    return calendar.get(java.util.Calendar.MONTH) + 1;
-  }
-  
-   /**
-   * Helper function to know the current year. 
-   *
-   * @param calendar an instance of Calendar, so can know the current date
-   * @return an int representing the current year
-   */
-  public static int getCurrentYear(java.util.Calendar calendar) {
-    return calendar.get(java.util.Calendar.YEAR);
-  }
-  
-   /**
-   * Helper function to format the relevant times (start or end) for a course 
-   * being converted into a calendar event. 
-   *
-   * @param timeInMin the time, in minutes of the week, for an event (either the start or end time)
-   * @param month the current month
-   * @param year the current year
-   * @param nextSun the day of the month of the next Sunday
-   * 
-   * @return a String formatted for DateTime, to be used in creating a calendar event 
-   */
-  public static String calculateDateTime(int timeInMin, int month, int year, int nextSun) {
-    int dayOfWeek = (int) Math.floorDiv(timeInMin, 24 * 60);
-    int dayInHours = dayOfWeek * 24;
-    int hour = (int) Math.floor((timeInMin / 60) - dayInHours);
-    int minute = (int) (60 * (((timeInMin / 60.0) - dayInHours) - hour));
-    int day = dayOfWeek + nextSun;
-     
-    String dayString = day > 9 ? Integer.toString(day) : "0" + Integer.toString(day);
-    String monthString = month > 9 ? Integer.toString(month) : "0" + Integer.toString(month);
-    String hourString = hour > 9 ? Integer.toString(hour) : "0" + Integer.toString(hour);     
-    String minuteString = minute > 9 ? Integer.toString(minute) : "0" + Integer.toString(minute);
-
-    String dateTime = year + "-" + monthString + "-" + dayString + "T" + hourString + ":" + minuteString + ":00";
-    return dateTime;
-  }
-
-   /**
-   * Helper function to calculate the properly formatted start time of a course
-   *
-   * @param timerange the time range for a course's meeting time (ex. 10:30-12:00)
-   * @param month the current month
-   * @param year the current year
-   * @param nextSun the day of the month of the next Sunday
-   * 
-   * @return a string representing the start time
-   */
-  public static String calculateStartTime(TimeRange timerange, int month, int year, int nextSun) {
-    // user will input their semester / quarter start date
-    int startTimeInMin = timerange.start();
-    String startTime = calculateDateTime(startTimeInMin, month, year, nextSun);
-    return startTime;
-  }
-
-   /**
-   * Helper function to calculate the properly formatted end time of a course
-   *
-   * @param timerange the time range for a course's meeting time (ex. 10:30-12:00)
-   * @param month the current month
-   * @param year the current year
-   * @param nextSun the day of the month of the next Sunday
-   * 
-   * @return a string representing the end time
-   */
-  public static String calculateEndTime(TimeRange timerange, int month, int year, int nextSun) {
-    int endTimeInMin = timerange.end();
-    String endTime = calculateDateTime(endTimeInMin, month, year, nextSun);
-    return endTime;
-  }
-
-   /**
    * Helper function to add a course in the schedule to the calendar, as an event. 
    *
    * @param course the course being added 
@@ -202,8 +103,7 @@ public class ScheduleCalendar {
    * @param startTime the start time of the course
    * @param endTime the end time of the course
    */
-
-  public void addEvent(Course course, String startTime, String endTime) throws IOException {
+  private void addEvent(Course course, String startTime, String endTime) throws IOException {
     try {  
       /* Get an event from the primary calendar to get the timezone offset from the 
        event (to use for properly adjusting the start and end times to the correct timezone) */
